Move getstarted page inside app/add page
