+++ webapp/src/components/Layout.tsx
@@ -10,7 +10,6 @@ import {
 	MoonIcon,
 	PencilSquareIcon,
 	PuzzlePieceIcon,
-	QueueListIcon,
 	SunIcon,
 	UserGroupIcon,
 	WrenchScrewdriverIcon,
@@ -45,11 +44,6 @@ const noNavPages = [
 ];
 
 const agentNavigation: any[] = [
-	{
-		name: 'Get Started',
-		href: '/getstarted',
-		icon: <QueueListIcon className='h-6 w-6 shrink-0' aria-hidden='true' />
-	},
 	{
 		name: 'Apps',
 		href: '/apps',
